+++ src/components/Company/index.tsx
@@ -8,5 +8,5 @@ export { Locations } from './Locations/Locations'
 export { PaySchedule } from './PaySchedule'
 export { FederalTaxes } from './FederalTaxes'
 export { BankAccount } from './BankAccount/BankAccount'
-export { StateTaxesFlow } from './StateTaxes/StateTaxesFlow'
+export { StateTaxes } from './StateTaxes/StateTaxes'
 export { OnboardingFlow } from './OnboardingFlow/OnboardingFlow'
